+++ CHANGELOG.md
@@ -21,5 +21,8 @@ All Notable changes to `Parser` will be documented in this file
 
 ## Unreleased 0.1.0-alpha.2 [2016-05-25]
 
+### Added
+- #18 Pass request options to `get` and `post` methods
+
 ### Fixed
 - #16 Use `ElementFinderFactoryInterface` instead of `ElementFinderFactory`